adult_income/ml_process/requirements.txt
numpy==2.2.4
pandas==2.2.3
scikit-learn==1.6.1
scipy==1.15.2
joblib==1.4.2
mlflow==2.22.0
python-dotenv==1.1.0